+++ packages/cubejs-schema-compiler/src/adapter/PreAggregations.js
@@ -732,14 +732,13 @@ export class PreAggregations {
           this.query.aggregateOnGroupedColumn(
             measure.measureDefinition(),
             measure.aliasName(),
-            !this.query.safeEvaluateSymbolContext().overTimeSeriesAggregate
+            !this.query.safeEvaluateSymbolContext().overTimeSeriesAggregate,
+            path
           ) || `sum(${measure.aliasName()})`
         ];
       }),
       R.fromPairs
-    )(preAggregationForQuery.preAggregation.type === 'autoRollup' ?
-      preAggregationForQuery.preAggregation.measures :
-      this.evaluateAllReferences(preAggregationForQuery.cube, preAggregationForQuery.preAggregation).measures);
+    )(this.rollupMeasures(preAggregationForQuery));
 
     // TODO granularity shouldn't be null?
     const rollupGranularity = this.castGranularity(preAggregationForQuery.preAggregation.granularity) || 'day';
@@ -760,5 +759,11 @@ export class PreAggregations {
         rollupGranularity,
       }
     );
+  }
+
+  rollupMeasures(preAggregationForQuery) {
+    return preAggregationForQuery.preAggregation.type === 'autoRollup' ?
+      preAggregationForQuery.preAggregation.measures :
+      this.evaluateAllReferences(preAggregationForQuery.cube, preAggregationForQuery.preAggregation).measures;
   }
 }